backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-environ==0.11.2
Pillow==10.1.0
django-filter==23.3
python-dateutil==2.8.2
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3